move vendored functionalities to more descriptive modules
move the different functionalities vendored from older protobuf versions into
more descriptive modules within the util module.
