+++ compiler/src/util/mod.rs
@@ -16,7 +16,10 @@
 use std::fmt;
 use std::str;
 
-use crate::vendored::CodeWriter;
+pub mod scope;
+pub mod writer;
+
+use writer::CodeWriter;
 
 // A struct that divide a name into serveral parts that meets rust's guidelines.
 struct NameSpliter<'a> {
@@ -130,6 +133,15 @@ where
     F: Fn(&mut CodeWriter),
 {
     async_fn_block(w, false, sig, cb);
+}
+
+// proto_name_to_rs is constructor as "{proto_path_to_rust_mod}.rs"
+// see https://github.com/stepancheg/rust-protobuf/blob/v3.7.2/protobuf-codegen/src/gen/paths.rs#L43
+pub fn proto_path_to_rust_mod(path: &str) -> String {
+    protobuf_codegen::proto_name_to_rs(path)
+        .strip_suffix(".rs")
+        .unwrap()
+        .to_string()
 }
 
 pub enum MethodType {
